requirements.txt
Django>=6.0
djangorestframework>=3.14
pymysql>=1.1.0
